Set y, m, d as null in EDTFDate if not defined
Previously, checked if year, month, and day variables existed in object
creating a new instance of EDTFDate, when it's actually better to check
for falsey values instead. For example, we can now call EDTFDate with
something like {year: '1995', month: '9', day: false}, which, previously
would have created an EDTFDate with a day of string 'false' (bad).